Collateral in its own assembly
