+++ src/Stratis.Bitcoin.FullNode.sln
@@ -174,6 +174,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Stratis.Features.Diagnostic
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FodyNlogAdapter", "FodyNlogAdapter\FodyNlogAdapter.csproj", "{2576847F-BFFC-4F0A-B883-28F2E38FEC7F}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Stratis.Features.Collateral", "Stratis.Features.Collateral\Stratis.Features.Collateral.csproj", "{AFEA4601-7938-44D2-A295-DBA022069F5B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -452,6 +454,10 @@ Global
 		{2576847F-BFFC-4F0A-B883-28F2E38FEC7F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2576847F-BFFC-4F0A-B883-28F2E38FEC7F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2576847F-BFFC-4F0A-B883-28F2E38FEC7F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AFEA4601-7938-44D2-A295-DBA022069F5B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AFEA4601-7938-44D2-A295-DBA022069F5B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AFEA4601-7938-44D2-A295-DBA022069F5B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AFEA4601-7938-44D2-A295-DBA022069F5B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -515,6 +521,7 @@ Global
 		{C83A089D-7B30-46BB-BBE3-AFBBAFB9A40D} = {1B9A916F-DDAC-4675-B424-EDEDC1A58231}
 		{53209209-7239-4FEB-BE8D-B2DF6145BBDA} = {15D29FFD-6142-4DC5-AFFD-10BA0CA55C45}
 		{2576847F-BFFC-4F0A-B883-28F2E38FEC7F} = {1B9A916F-DDAC-4675-B424-EDEDC1A58231}
+		{AFEA4601-7938-44D2-A295-DBA022069F5B} = {15D29FFD-6142-4DC5-AFFD-10BA0CA55C45}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6C780ABA-5872-4B83-AD3F-A5BD423AD907}
