+++ src/Stratis.Features.Collateral/CheckCollateralFullValidationRule.cs
@@ -9,7 +9,7 @@ using Stratis.Bitcoin.Features.PoA;
 using Stratis.Bitcoin.Interfaces;
 using Stratis.Bitcoin.Utilities;
 
-namespace Stratis.Features.FederatedPeg.Collateral
+namespace Stratis.Features.Collateral
 {
     /// <summary>Ensures that collateral requirement on counterpart chain is fulfilled for the federation member that produced a block.</summary>
     /// <remarks>Ignored in IBD.</remarks>
