+++ src/Stratis.Features.Collateral/CounterChain/CounterChainSettings.cs
@@ -1,9 +1,8 @@
 ﻿using NBitcoin;
 using Stratis.Bitcoin.Configuration;
 using Stratis.Bitcoin.Utilities;
-using Stratis.Features.FederatedPeg.Interfaces;
 
-namespace Stratis.Features.FederatedPeg.CounterChain
+namespace Stratis.Features.Collateral.CounterChain
 {
     public class CounterChainSettings : ICounterChainSettings
     {
