+++ src/Stratis.Features.FederatedPeg.IntegrationTests/Utils/SidechainTestContext.cs
@@ -16,7 +16,7 @@ using Stratis.Bitcoin.IntegrationTests.Common;
 using Stratis.Bitcoin.IntegrationTests.Common.EnvironmentMockUpHelpers;
 using Stratis.Bitcoin.Networks;
 using Stratis.Bitcoin.Tests.Common;
-using Stratis.Features.FederatedPeg.CounterChain;
+using Stratis.Features.Collateral.CounterChain;
 using Stratis.Features.FederatedPeg.Models;
 using Stratis.Sidechains.Networks;
