+++ src/Stratis.Features.FederatedPeg/Controllers/FederationGatewayClient.cs
@@ -4,7 +4,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Stratis.Bitcoin.Controllers;
-using Stratis.Features.FederatedPeg.Interfaces;
+using Stratis.Features.Collateral.CounterChain;
 using Stratis.Features.FederatedPeg.Models;
 
 namespace Stratis.Features.FederatedPeg.Controllers
