+++ src/Stratis.Features.FederatedPeg/OpReturnDataReader.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using Microsoft.Extensions.Logging;
 using NBitcoin;
-using Stratis.Features.FederatedPeg.CounterChain;
+using Stratis.Features.Collateral.CounterChain;
 using TracerAttributes;
 
 namespace Stratis.Features.FederatedPeg
