+++ src/Stratis.Features.FederatedPeg/Stratis.Features.FederatedPeg.csproj
@@ -25,6 +25,7 @@
     <ProjectReference Include="..\Stratis.Bitcoin.Features.Api\Stratis.Bitcoin.Features.Api.csproj" />
     <ProjectReference Include="..\Stratis.Bitcoin.Features.Notifications\Stratis.Bitcoin.Features.Notifications.csproj" />
     <ProjectReference Include="..\Stratis.Bitcoin.Features.SmartContracts\Stratis.Bitcoin.Features.SmartContracts.csproj" />
+    <ProjectReference Include="..\Stratis.Features.Collateral\Stratis.Features.Collateral.csproj" />
   </ItemGroup>
 
 </Project>
